feat: 결제금액 허용 범위 long 범위로 확대

+++ src/main/java/roomescape/reservation/dto/request/PaymentRequest.java
@@ -1,12 +1,12 @@
 package roomescape.reservation.dto.request;
 
 public record PaymentRequest(
-        int amount,
+        long amount,
         String orderId,
         String paymentKey
 ) {
 
-    public static PaymentRequest toRequest(ReservationCreateRequest reservationCreateRequest){
+    public static PaymentRequest toRequest(ReservationCreateRequest reservationCreateRequest) {
         return new PaymentRequest(reservationCreateRequest.amount(), reservationCreateRequest.orderId(),
                 reservationCreateRequest.paymentKey());
     }
